sidewalk v0.3.1 Sidewalk.Job

This struct represents a typical Sidewalk job.

Struct definitions

jid -> A 24 character long job identifier

class -> The worker class which is responsible for executing the job.

args -> The arguments passed which should be passed to the worker.

created_at -> The timestamp when the job is created by Sidewalk.

enqueue_at -> The timestamp when the job is really enqueued with the Redis server.

queue -> The queue where a job should be enqueued. Defaults to “default” queue.

retry -> Tells the Sidekiq worker to retry the enqueue job.

Types

t()

t() :: %Sidewalk.Job{args: List.t, class: String.t, created_at: Float.t, enqueued_at: Float.t, jid: String.t, queue: String.t, retry: boolean}